+++ examples/positioning_elements/app.tsx
@@ -3,6 +3,7 @@ import {
   Button,
   FormField,
   Grid,
+  ImageCard,
   Rows,
   Select,
   Text,
@@ -16,10 +17,8 @@ import {
 import cat from "assets/images/cat.jpg";
 import dog from "assets/images/dog.jpg";
 import rabbit from "assets/images/rabbit.jpg";
-import clsx from "clsx";
 import React from "react";
 import baseStyles from "styles/components.css";
-import styles from "./styles.css";
 
 // Below values are only for demonstration purposes.0
 // You can position your elements anywhere on the page by providing arbitrary
@@ -183,15 +182,14 @@ export const App = () => {
             <Box id={props.id} padding="1u">
               <Grid columns={3} spacing="1.5u">
                 {items.map((item) => (
-                  <img
-                    className={clsx(
-                      styles.thumbnail,
-                      item.active && styles.active
-                    )}
+                  <ImageCard
+                    ariaLabel={item.title}
                     key={item.key}
-                    src={item.imageSrc}
+                    thumbnailUrl={item.imageSrc}
                     onClick={item.onClick}
-                    alt={item.title}
+                    selectable={true}
+                    selected={item.active}
+                    borderRadius="standard"
                   />
                 ))}
               </Grid>